move: FSD 구조에 맞게 파일 이동 (#107)
