+++ apps/web/src/app/(with-local-layout)/(with-header-layout)/auctions/page.tsx
@@ -1,10 +1,10 @@
 //TODO - 하단의 경매 현황 누르면 url 파라미터로 end_time을 넘기도록 해야 함
 
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
-import AuctionList from 'src/components/auctions/AuctionList';
-import SelectOrder from 'src/components/auctions/SelectOrder';
-import PageContainer from 'src/components/layout/PageContainer';
 import { fetchAllAuctionWithEpisodeCount } from 'src/entities/auction/api';
+import AuctionList from 'src/features/auction/AuctionList';
+import SelectOrder from 'src/features/auction/SelectOrder';
+import PageContainer from 'src/shared/PageContainer';
 import { AuctionRow } from 'src/shared/supabase/types';
 
 interface EpisodeCount {
