+++ apps/web/src/features/auction/shared/AuctionCard.tsx
